requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cx-Oracle==8.2.1
Django==3.1.5
django-partial-date==1.3.1
django-rest==0.8.6
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.9
Pillow==8.1.0
psycopg2-binary==2.8.6
Pygments==2.7.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
